==> Data-Analysis-Viz/src/0--Install_Libraries.ipynb <==
# Install Libraries

**This notebook installs libraries that are not already present in the Jupyter environment provided for SIADS-593 via Coursera.**

## pyreadstat

This module is required to convert the BRFSS file from SAS (XPT) format to CSV

!pip install pyreadstat

## watermark

This module provides a watermark functionality for all notebooks to record dependencies

!pip install watermark

## us

This module makes it easier to work with U.S. metadata 

!pip install us

## geopandas

This library extends Pandas by adding support for geospatial data

!pip install geopandas
