==> Data-Analysis-Viz/src/5-V-Merge_datasets.ipynb <==
# Merge the two datasets

_Primary Author: Vikram Vadammani_

import pandas as pd
import numpy as np
import scipy.stats as sps
from scipy.stats import zscore
from sklearn.preprocessing import MinMaxScaler

PATH = 'data/'

brfssdf = pd.read_csv(PATH + 'brfss_2022.csv')
#brfssdf = brfssdf.head(1000) # limit the df so we can load it easier

# not limiting the df with head or dropna cus we need to see the whole df to make sure the state merge is working. Also dropna kind of drops all the data.
# This means that there are lots of NaNs in the data
print(f'Total rows in BRFSS: {len(brfssdf)}')
brfssdf.head()

# Filter BRFSS dataset to include only 2022
brfssdf = brfssdf[brfssdf["IYEAR"] == 2022].copy()

# Drop the "IYEAR" column from BRFSS after filtering (optional)
brfssdf.drop(columns=["IYEAR"], inplace=True)

#going to find all of the cols that are in anyway related to depression
#A lot of these collumns aent going into the scoring system but it might be nice to see the data of these things AFTER we use the scoring system

#Descriptions of variables  vvvv

cols = ['MENTHLTH','ADDEPEV3', 'SLEPTIM1','EXERANY2','AVEDRNK3','LCSNUMCG', 'ACEDEPRS','LSATISFY','EMTSUPRT', 'SDHISOLT', 'SDHSTRE1','MARIJAN1','POORHLTH']
vardesc = {'MENTHLTH': 'number of bad mental days', 'ADDEPEV3': 'had a depressive episode?', 'SLEPTIM1':'how much sleep', 'EXERANY2': 'exercise in past 30day?', 'AVEDRNK3': 'avg drink past 30day', 'LCSNUMCG': 'number of cigs past 30day', 'ACEDEPRS': 'live with depressed,ill or suici?', 'LSATISFY': 'satisfaction in life?', 'EMTSUPRT': 'how often do you get emo support?', 'SDHISOLT': 'how often socially isolated feeling ', 'SDHSTRE1': 'how often stressed', 'MARIJAN1': 'past 30day weed ', 'POORHLTH': 'number of poor mental or physical days past 30 da'}

# Keep these columns cus they are supported by the PHQ-9 survey AND they have good data in the BRFSS set
scoringcols = ['_STATE','MENTHLTH','ADDEPEV3']
scoringdf = brfssdf[scoringcols].copy()

# Show missing values
missing_values = scoringdf.isna().sum()
print('Missing Values Before Processing:')
print(missing_values[missing_values > 0])


#now I want to convert the _STATE column from floats to strings

state_mapping = {
    1: 'AL', 2: 'AK', 4: 'AZ', 5: 'AR', 6: 'CA',
    8: 'CO', 9: 'CT', 10: 'DE', 11: 'DC', 12: 'FL', 13: 'GA',
    15: 'HI', 16: 'ID', 17: 'IL', 18: 'IN', 19: 'IA', 20: 'KS', 21: 'KY', 22: 'LA',
    23: 'ME', 24: 'MD', 25: 'MA', 26: 'MI', 27: 'MN', 28: 'MS', 29: 'MO', 30: 'MT',
    31: 'NE', 32: 'NV', 33: 'NH', 34: 'NJ', 35: 'NM', 36: 'NY', 37: 'NC', 38: 'ND',
    39: 'OH', 40: 'OK', 41: 'OR', 42: 'PA', 44: 'RI', 45: 'SC', 46: 'SD', 47: 'TN',
    48: 'TX', 49: 'UT', 50: 'VT', 51: 'VA', 53: 'WA', 54: 'WV', 55: 'WI', 56: 'WY',
    66: 'GU', 72: 'PR', 78: 'VI'  # Territories: Guam, Puerto Rico, Virgin Islands
}

# Convert float state codes to abbreviations
scoringdf['_STATE'] = scoringdf['_STATE'].map(state_mapping)

#now I want to only keep the states that are also in the daylight states
valid_states = {
    'AL', 'AK', 'AZ', 'AR', 'CA', 'CO', 'CT', 'DE', 'FL', 'GA',
    'HI', 'ID', 'IL', 'IN', 'IA', 'KS', 'KY', 'LA', 'ME', 'MD',
    'MA', 'MI', 'MN', 'MS', 'MO', 'MT', 'NE', 'NV', 'NH', 'NJ',
    'NM', 'NY', 'NC', 'ND', 'OH', 'OK', 'OR', 'PA', 'RI', 'SC',
    'SD', 'TN', 'TX', 'UT', 'VT', 'VA', 'WA', 'WV', 'WI', 'WY',
    'PR'  # Puerto Rico
}

scoringdf = (scoringdf[scoringdf['_STATE'].isin(valid_states)]).dropna()
scoringdf.head()

scoringdf.describe()

#goal here is to normalize the values and add a Depres Index column

#there are values in the columns that are place holder values that represent something else ie. refused to answer, None, didnt answer.
menthlth_invalid = {77.0, 99.0}
addepev3_invalid = {7.0, 9.0}

#get rid of invalid values for the 2 columns
scoringdf = scoringdf[~scoringdf['MENTHLTH'].isin(menthlth_invalid) & ~scoringdf['ADDEPEV3'].isin(addepev3_invalid)].copy()

# Convert ADDEPEV3 (Depression Diagnosis) into binary format
scoringdf["ADDEPEV3"] = scoringdf["ADDEPEV3"].map({1.0: 1.0, 2.0: 0.0})

# Convert special code 88 for MENTHLTH to reflect zero days
MENTHLTH_map = {1.:1., 2.:2., 3.:3., 4.:4., 5.:5., 6.:6., 7.:7., 8.:8.,
                9.:9., 10.:10., 11.:11., 12.:12., 13.:13., 14.:14., 15.:15.,
                16.:16, 17.:17., 18.:18., 19.:19., 20.:20., 21.:21., 22.:22.,
                23.:23., 24.:24., 25.:25., 26.:26., 27.:27., 28.:28., 29.:29.,
                30.:30., 88:0.}

scoringdf['MENTHLTH'] = scoringdf['MENTHLTH'].map(MENTHLTH_map)

#Verify data
print(f'Unique values for ADDEPEV3: {scoringdf.ADDEPEV3.unique()}')
print(f'Unique values for MENTHLTH: {sorted(scoringdf.MENTHLTH.unique())}')
scoringdf.head()

scoringdf.describe()

#then we want to scale the number of bad mental health days between 0 and 1
#so that it is on the same scale as the ADDEPEV3
scaleddf = scoringdf.copy()
scaler1 = MinMaxScaler()
scaleddf['MENTHLTH'] = scaler1.fit_transform(scaleddf['MENTHLTH'].values.reshape(-1, 1))

#now apply the weights that we set for our Depression Index
scaleddf['ADDEPEV3'] = scaleddf['ADDEPEV3'] * 10
scaleddf['MENTHLTH'] = scaleddf['MENTHLTH'] * 8

#make a column for the DI that is both columns added together, on a 0-10 scale.
scaler2 = MinMaxScaler(feature_range=(0,10))
scaleddf['DI'] = scaler2.fit_transform(
    (scaleddf['MENTHLTH'] + scaleddf['ADDEPEV3']).values.reshape(-1, 1)
    )
scaleddf.head()

# Read transformed daylight notebook in from disk
daylight_trans = pd.read_csv(PATH + 'daylight_transformed.csv')
daylight_trans.head()

# Merge transformed daylight data and transformed BRFSS data
analysis_df = pd.merge(daylight_trans[['State', 'DH']],
                       scaleddf[['_STATE', 'DI', 'ADDEPEV3', 'MENTHLTH']],
                       left_on='State', right_on='_STATE', how='inner'
                      ).drop('_STATE', axis=1)
analysis_df.head()

# Normalize the 'Daylight Hours' column (Z-score)
analysis_df['DH_Z'] = sps.zscore(analysis_df.DH.values)

# Normalize the Depression Index (Z-score)
analysis_df['DI_Z'] = zscore(analysis_df.DI.values)
analysis_df.describe()

Note how the Z-score transforms each data column so that it has zero mean and unit variance.

# Write merged and transformed data frame to disk for collaborative notebooks
analysis_df.to_csv(PATH + 'analysis_2022.csv', index=False)

## Record Dependencies

%load_ext watermark
%watermark
%watermark --iversions
