==> Data-Analysis-Viz/src/7-K-Combined_Analysis.ipynb <==
#  OLS Regression Analysis

_Primary Author: Pete King_

This notebook uses the statsmodels library to perform an ordinary least squares (OLS) regression on the data, looking for a correlation between daylight and depression.  We also calculate the Pearson correlation coefficient, which is limited in range between -1 (strong negative correlation) and +1 (strong positive correlation).

import numpy as np
import pandas as pd
import scipy.stats as sps
import statsmodels.api as sm

PATH = 'data/'

analysis_df = pd.read_csv(PATH + 'analysis_2022.csv')
analysis_df.head()

# Note how both variables now have zero mean and unit variance.
analysis_df.describe()

# Prep data for an OLS Regression using the statsmodels library
# Add a column of 1's for the intercept calculation:
X = sm.add_constant(analysis_df.DH.values)
y = analysis_df.DI.values

# Perform OLS Regression and look at the summary table
model = sm.OLS(y, X)
results = model.fit()
summary = results.summary(xname=['intercept', 'Daylight Hours'], yname='Depression Index (0-10)')
print(summary)

# Calculate Pearson's correlation coefficient
pearson = sps.pearsonr(analysis_df.DH.values, analysis_df.DI.values)
r, p_value = pearson.statistic, pearson.pvalue
conf_int = pearson.confidence_interval(0.95)
print(f'Pearson R: {r}')
print(f'p-value: {p_value}')
print(conf_int)

## Interpreting the Results

_For a detailed writeup, see the final report._

In summary, the t-statistic of -4.142 with a p-value of 0.000 provides strong evidence of a statistically significant correlation between our two variables of interest.  We cannot assume a causal relationship, but if we did assume one, we could say that the magnitude of the effect from 'daylight hours' on our 'depression index' is extremely small, as evidenced both by the R-squared value of 0.000 and a Pearson correlation coefficient of -0.0065.

What does this mean, practically speaking?  It means that, while there does appear to be a correlation between daylight and depression, the 'effect' is minimal.  We could say that for every additional hour of daylight available to a person, we can expect their depression index score (on a scale of 0-10) to decrease by 0.03, on average.  Other factors appear to play a much larger role in determining the degree to which a person feels depressed.

## Record Dependencies

%load_ext watermark
%watermark
%watermark --iversions
